Directories and profiles + code refactoring
